check values of efeature matrix
