+++ featuretools/tests/computational_backend/test_calculate_feature_matrix.py
@@ -13,7 +13,6 @@ import pytest
 from dask import dataframe as dd
 from tqdm import tqdm
 from woodwork.column_schema import ColumnSchema
-from woodwork.logical_types import Integer
 
 import featuretools as ft
 from featuretools import EntitySet, Timedelta, calculate_feature_matrix, dfs
@@ -38,7 +37,6 @@ from featuretools.primitives import (
     Count,
     Max,
     Min,
-    NumericLag,
     Percentile,
     Sum,
     TransformPrimitive
@@ -1870,34 +1868,3 @@ def test_cfm_with_invalid_time_index(es):
     match += "which differs from other entityset time indexes"
     with pytest.raises(TypeError, match=match):
         calculate_feature_matrix(features=features, entityset=es)
-
-
-def test_cfm_with_lag_and_non_nullable_column(pd_es):
-    # fill nans so we can use non nullable numeric logical type in the EntitySet
-    new_log = pd_es['log'].copy()
-    new_log['value'] = new_log['value'].fillna(0)
-    new_log.ww.init(logical_types={'value': 'Integer',
-                                   "product_id": "Categorical"},
-                    index='id', time_index='datetime',
-                    name='new_log')
-    pd_es.add_dataframe(new_log)
-    rels = [('sessions', 'id', 'new_log', 'session_id'),
-            ('products', 'id', 'new_log', 'product_id')]
-    pd_es = pd_es.add_relationships(rels)
-
-    assert isinstance(pd_es['new_log'].ww.logical_types['value'], Integer)
-
-    lag_primitive = NumericLag(periods=5)
-    features = ft.dfs(target_dataframe_name='new_log',
-                      entityset=pd_es,
-                      agg_primitives=[],
-                      trans_primitives=[lag_primitive],
-                      features_only=True)
-
-    fm = calculate_feature_matrix(features=features, entityset=pd_es)
-
-    # Non nullable
-    assert "NUMERIC_LAG(datetime, value, periods=5)" in fm.columns
-    # Nullable
-    assert "NUMERIC_LAG(datetime, value_2, periods=5)" in fm.columns
-    assert "NUMERIC_LAG(datetime, products.rating, periods=5)" in fm.columns
