+++ featuretools/tests/primitive_tests/test_transform_features.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 import pytest
 from woodwork.column_schema import ColumnSchema
-from woodwork.logical_types import Boolean, Datetime
+from woodwork.logical_types import Boolean, Datetime, Integer
 
 import featuretools as ft
 from featuretools.computational_backends.feature_set import FeatureSet
@@ -44,6 +44,7 @@ from featuretools.primitives import (
     NotEqual,
     NotEqualScalar,
     NumCharacters,
+    NumericLag,
     NumWords,
     Percentile,
     ScalarSubtractNumericFeature,
@@ -1113,3 +1114,35 @@ def test_time_since_primitive_matches_all_datetime_types(es):
 
     for name in expected_names:
         assert name in fm.columns
+
+
+def test_cfm_with_lag_and_non_nullable_column(pd_es):
+    # fill nans so we can use non nullable numeric logical type in the EntitySet
+    new_log = pd_es['log'].copy()
+    new_log['value'] = new_log['value'].fillna(0)
+    new_log.ww.init(logical_types={'value': 'Integer',
+                                   "product_id": "Categorical"},
+                    index='id', time_index='datetime',
+                    name='new_log')
+    pd_es.add_dataframe(new_log)
+    rels = [('sessions', 'id', 'new_log', 'session_id'),
+            ('products', 'id', 'new_log', 'product_id')]
+    pd_es = pd_es.add_relationships(rels)
+
+    assert isinstance(pd_es['new_log'].ww.logical_types['value'], Integer)
+
+    lag_primitive = NumericLag(periods=5)
+    fm, _ = ft.dfs(target_dataframe_name='new_log',
+                   entityset=pd_es,
+                   agg_primitives=[],
+                   trans_primitives=[lag_primitive])
+
+    # Non nullable
+    assert fm["NUMERIC_LAG(datetime, value, periods=5)"].head(5).isnull().all()
+    assert fm["NUMERIC_LAG(datetime, value, periods=5)"].isnull().sum() == 5
+    # Nullable
+    assert "NUMERIC_LAG(datetime, value_2, periods=5)" in fm.columns
+    assert fm["NUMERIC_LAG(datetime, products.rating, periods=5)"].head(5).isnull().all()
+
+    assert "NUMERIC_LAG(datetime, products.rating, periods=5)" in fm.columns
+    assert fm["NUMERIC_LAG(datetime, products.rating, periods=5)"].head(5).isnull().all()
